# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Simple API",
        "description": "A simple API to learn how to write OpenAPI Specification"
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
...
        "/api/posts/": {
            "post": {
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Body Object",
                        "required": false,
                        "schema": {
                            "title": "Post",
                            "required": [
                                "text"
                            ],
                            "properties": {
                                "user": {
                                    "type": "string"
                                },
                                "text": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "avatar": {
                                    "type": "string"
                                },
                                "likes": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "user": {
                                                "type": "string"
                                            }
                                        },
                                        "required": []
                                    }
                                },
                                "comments": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "user": {
                                                "type": "string"
                                            },
                                            "text": {
                                                "type": "string"
                                            },
                                            "avatar": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "required": [
                                            "text"
                                        ]
                                    }
                                },
                                "date": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "get": {
                "parameters": [],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "title": "Post",
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "user": {
                                        "type": "string"
                                    },
                                    "text": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "type": "string"
                                    },
                                    "likes": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "user": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": []
                                        }
                                    },
                                    "comments": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "user": {
                                                    "type": "string"
                                                },
                                                "text": {
                                                    "type": "string"
                                                },
                                                "avatar": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "required": [
                                                "text"
                                            ]
                                        }
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{post_id}/": {
            "get": {
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "search string",
                        "required": true,
                        "type": "string"
                    }
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "search string",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/api/posts/like/{id}/": {
            "put": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search string",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/api/posts/unlike/{id}/": {
            "put": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search string",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/api/posts/comment/{id}/": {
            "post": {
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search string",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
        "/api/posts/comment/{post_id}/{comment_id}/": {
            "delete": {
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "search string",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "search string",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
...
    }
}